example/src/openai_agent_example/requirements.txt
fastapi
uvicorn
pydantic
openai
openai-agents
python-dotenv
boto3 